+++ src/pages/LandingPage/Sections/FeatureFocusSection.jsx
@@ -1,0 +1,21 @@
+export default function FeatureFocusSection() {
+  return (
+    <section className="feature-focus">
+      <div className="container mx-auto text-center">
+        <div className="p-20">
+          <h2 className="mb-4 text-5xl font-bold leading-10 tracking-tight text-[#4a131a]">
+            Your Stories, Our Progress
+          </h2>
+          <div className="text-lg leading-loose text-[#952735]">
+            Your voice contributes to a data-driven transformation that benefits the entire diving
+            community.
+            <p>
+              Be a driving force behind safer diving practices, community development, and lives
+              saved.
+            </p>
+          </div>
+        </div>
+      </div>
+    </section>
+  );
+}
