+++ src/pages/LandingPage/Sections/index.js
@@ -1,0 +1,5 @@
+export { default as HeroSection } from './HeroSection';
+export { default as HighlightsSection } from './HighlightsSection';
+export { default as FeatureFocusSection } from './FeatureFocusSection';
+export { default as CTASection } from './CTASection';
+export { FeatureContentLeftSection, FeatureContentRightSection } from './FeatureContentSection';
